refactor: 외부 서버 exception추가, 강수량 타입 변경(Double)

+++ src/main/java/com/techeer/fashioncloud/domain/weather/dto/UltraSrtNcstResponse.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class UltraSrtNcstResponse implements ForecastResponse {
 
     private Double temperature;
-    private Integer hourRainfall;
+    private Double hourRainfall;
     private Integer humidity;
     private Integer rainfallType;
     private Double windSpeed;
